### 2017-02-16 #99 @sckott
cmipr - Client for Coupled Model Intercomparison Project (CMIP) data
Client to work with CMIP data - downscaled climate and hydrology projections. Package lists avail. files, downloads and caches, and reads into raster objects.

```
Package: cmipr
Title: Client for Coupled Model Intercomparison Project (CMIP) Data
Description: Client for Coupled Model Intercomparison Project (CMIP) data
		(<http://gdo-dcp.ucllnl.org/downscaled_cmip_projections/>).
		Data is stored on an FTP server, from which we provide
		functions to fetch data and return tidy data.
Version: 0.0.8.9310
Authors@R: person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com")
    hoardr (>= 0.0.3),
    raster (>= 2.5-8),
    ncdf4 (>= 1.15),
    curl (>= 2.3),
    crul (>= 0.2.1.9100),
    tibble
    roxygen2 (>= 6.0.1),
    covr
Remotes: ropensci/crul, ropensci/hoardr
```

https://github.com/ropenscilabs/cmipr

People using climate projections for research - or studying climate change itself.

https://github.com/JGCRI/RCMIP5 deals with CMIP data, but AFAICT only handles data after you have it on your machine. So that package and this one could be used together.

maybe https://github.com/bpbond since author of the above mentioned pkg
