### 2017-02-17 #99 @masalmon
### Editor checks:

---

#### Editor comments

Thank you for your submission @sckott!

* RCMIP5 doc states "This package does not handle downloading (i.e. from nodes in the Earth System Grid Federation, http://esgf.org) the NetCDF data themselves. Sorry."
So AFAICT too there's no overlap.

* Here is the output from `goodpractice::gp()`.

```r
It is good practice to

  ? write unit tests for all functions, and all package code in general.
    80% of code lines are covered by test cases.

    R/cmip_fetch.R:24:NA
    R/cmip_fetch.R:25:NA
    R/cmip_fetch.R:26:NA
    R/cmip_read.R:41:NA
    R/cmip_read.R:43:NA
    ... and 8 more lines

  ? avoid long code lines, it is bad for readability. Also, many people
    prefer editor windows that are about 80 characters wide. Try make your
    lines shorter than 80 characters

    tests\testthat\test-cmip_fetch.R:4:1
    tests\testthat\test-cmip_read.R:4:1
```

where the long lines are only due to paths so you could only add "# nolint" at the end of these lines. :wink:

* I identified one typo using `devtools::spell_check()`. "Reaa" in cmip_read.Rd:5,17 should be "Read".

---

Currently seeking reviewers

Reviewers:
Due date:
