### 2017-02-17 #99 @sckott
thanks @masalmon will fix those things
