### 2017-02-17 #99 @sckott
@masalmon okay, those 2 thigns fixed
