### 2017-02-19 #99 @masalmon
Thanks for accepting to review this package @cvitolo and @bpbond !

As a reminder here are links to the recently updated [reviewing](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and [packaging](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md) guides and to the [review template](https://raw.githubusercontent.com/ropensci/onboarding/master/reviewer_template.md).
