### 2017-03-11 ropenscilabs/cmipr#10 @cvitolo
Non descriptive caption in vignette
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

The plot at the end of the vignette has the following caption 'plot of chunk unnamed-chunk-16', it would be more informative to have a description of the figure.
