### 2017-03-11 ropenscilabs/cmipr#11 @cvitolo
Content of the inst folder
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

I noticed the inst folder seems to contain redundant files, could it be deleted?
