### 2017-03-11 ropenscilabs/cmipr#12 @cvitolo
More info needed in function documentation: cmip_list_files()
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

The function `cmip_list_files` could contain a brief description of:

- the available options for the `route` argument and
- the expected content of the output table (what is the `date` column? what is the `file` column?).
