### 2017-03-11 ropenscilabs/cmipr#13 @cvitolo
cmip_list_files() returns empty records under the `file` column (only lists the folders under bcsd)
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

I thought `cmip_list_files()` would return all the 1st level folders in the root directory of the ftp server like in the screenshot below:

![screenshot from 2017-03-11 21 22 02](https://cloud.githubusercontent.com/assets/4742575/23826983/f5221b56-06a0-11e7-970c-46a0a295a8e5.png)

However, if I run `cmip_list_files()` the output table only lists the folders under bcsd. Also there are some missing entries in the `file` column and I'm not sure of what that means. Is this a bug?
