### 2017-03-11 ropenscilabs/cmipr#14 @cvitolo
cmip_fetch(), route argument documentation
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

The documentation for `cmip_fetch` says that the route argument is the 'Route to a folder, a character string', but in the example the key contains the path to a (netcdf) file, not a folder.  The description of the route argument should be amended.

Can you fetch the content of an entire folder? If so, maybe you could add another example.
