### 2017-03-11 ropenscilabs/cmipr#15 @cvitolo
cmip_fetch() with `overwrite = TRUE` is not re-fetching the data
I tried to run the following

`key <- "bcsd/yearly/cnrm_cm3.1/cnrm_cm3.1.sresa1b.monthly.Prcp.2034.nc"`
`res <- cmip_fetch(key)`
`res1 <- cmip_fetch(key, overwrite = TRUE)`

But noticed that the second time, it is not re-fetching the data. Is this the expected behaviour?
