### 2017-03-11 ropenscilabs/cmipr#16 @cvitolo
NetCDF library as dependency
This issue is part of [this rOpenSci review](https://github.com/ropensci/onboarding/issues/99).

I think that the README should mention the NetCDF library as external dependency, without this library the package won't install.
