### 2017-03-11 #99 @cvitolo
I think that the readme should mention the netcdf library as external dependency.

---

@masalmon @sckott in my opinion this package is well written and it does what it says. I only have very few comments which I filed as issues, also reporting specific examples.

The main comment is that all the functions contain basic documentation but it would be great to have some more information. Once there is a more in-depth documentation, you could use `pkgdown` to automatically create a website for your package (or alternatively create a github-page). @masalmon suggested that for one of my packages not long ago and I think it's a great idea because your users might find that more visually appealing than the standard R help pages.

Also, please consider mentioning the NetCDF library as external dependency as ncdf4 is required by the raster package to handle nc files in your examples.
