### 2017-03-13 #99 @masalmon
Also @cvitolo the different issues are fine now but note for future reviews that for rOpenSci's reviews all is usually kept in one thread, not like what happens at JOSS (rOpenSci package authors sometimes open issues themselves when dealing with the review  as you see in https://github.com/ropensci/onboarding/issues/94 ). 😊
