### 2017-03-13 #99 @cvitolo
@masalmon @sckott my apologies!
When I had my packages reviewed, I found very convenient to work on issues because they can develop in separate threads. This is extremely useful when the issue is complex and/or has some unexpected consequences on the remaining of the code.
Anyway, I'll follow best practice next time :)
