### 2017-03-13 ropenscilabs/cmip#15 @bpbond
Yes. I also couldn't get the `"file exists and ovewrite != TRUE"` message to print, so probably look hard at this logic/documentation.
