### 2017-03-13 ropenscilabs/cmip#15 @bpbond
Note also there's no test that checks whether `overwrite` works or not.
