### 2017-03-13 #99 @bpbond
---

This package provides an interface to download (a subset of) Coupled Model Intercomparison Project (CMIP, see [Taylor et al. 2013](http://journals.ametsoc.org/doi/abs/10.1175/BAMS-D-11-00094.1)) files from `ftp://gdo-dcp.ucllnl.org/pub/dcp/archive/`. This could be quite useful for R users interested in a scriptable way to download such files.

Overall, this package is well designed and well structured. The R code is clean and clearly written and formatted, although a bit light on comments. Test coverage is good (although see https://github.com/ropenscilabs/cmipr/issues/15). In generally the package complies with the ROpenSci and Mozilla guides. Documentation is thin--especially the function descriptions, which are extremely short and should be fleshed out--but what is there is clear and well written. R CMD CHECK passes with no errors, warnings, or notes.

A few suggestions:

* Would be useful if `cmip_list_files` listed file sizes along with date and filename
* `cmip_fetch` overwrite parameter doesn't seem to work?
* An optional progress bar during fetching might be useful.
* There's no `CONTRIBUTING.md` included.

**The biggest problem** is that the package needs to make clear it's using `ftp://gdo-dcp.ucllnl.org/pub/dcp/archive/`, which as far as I can see provides only a tiny subset of CMIP data--only temperature and precipitation? I'm concerned that people will be confused that they can't download, e.g., ocean pH data. This also means that the package is vulnerable if LLNL's ftp service is disrupted or changes, as there's no failover or backup source of data. Note that there are much more complete CMIP archives available online (e.g. the `atmos.ethz.ch` rsync server, although this is a private server and you'd need to contact them for access).

My test environment:
```
Platform: x86_64-apple-darwin13.4.0 (64-bit)
Running under: OS X El Capitan 10.11.6
```
