### 2017-03-14 #99 @sckott
thanks for your review @bpbond !

So I can get a complete view of CMIP data:

- I'm getting data in this pkg from the ftp server
- There's more complete data available via rsync from `atmos.ehz.ch`
- Are there any other places to get CMIP data?

Do you think the folks behind `atmos.ehz.ch`  would allow public access - or is it a fill out a form, wait for email, kind of thing?
