### 2017-03-14 #99 @bpbond
Other places to get data - well, the canonical 'place' is the Earth System Grid Federation but that's been a source of huge frustration for _lots_ of people (which is probably why LLNL established its ftp server). I'm sure there are other private repositories too (we have one, but it's sealed behind a firewall that I can't control).

The `atmos.ehz.ch` server is amazing, a huge data store. I have no idea if they'd allow public access or not, but it doesn't hurt to ask. The contact I have for that server (hopefully not outdated) is [Urs Beyerle](http://www.iac.ethz.ch/people-iac/person-detail.html?persid=49918).
