### 2017-03-20 #99 @sckott
Was told that public access not possible - but pointed me towards https://github.com/Prodiguer/synda - that not useable here, but sounds like there might be some open web services that synda uses that we can use, e.g., at this location http://esgf-index1.ceda.ac.uk/esg-search/search
