### 2017-03-22 #99 @sckott
@bpbond heard back from ATMOS folks

have you ever used the Synda command line tool? https://github.com/Prodiguer/synda They suggested to have users use that on the command line and then just read data in within R - as opposed to doing the downloads ourselves here. thoughts?

He said:

> Have you ever looked into original CMIP5 data? It can be quite a mess. The archive you mentioned (ftp://gdo-dcp.ucllnl.org/pub/dcp/archive/cmip5/) looks already
like a processed data archive (http://gdo-dcp.ucllnl.org). This data is in much better shape than the original CMIP5 data. And therefore much easier to process.

may be true, but you said it's only partial data, yes?
