### 2017-03-23 #99 @bpbond
Hey Scott. No, I don't know Synda–huh, interesting, thanks.

I agree that the original CMIP5 data can be a mess...but yeah, LLNL seems (as far as I could see) to be _extremely_ partial: as in, out of the hundreds of CMIP5 output variables, they only seem to have two or three.

`cmipr` would still bring value if it just handles downloads from there, I guess, but it's pretty limited, and so as to not waste folks' time you'll want to make it extremely clear what's available and what's not. Bummer.
