### 2017-03-23 #99 @sckott
@bpbond I think we really just need to re-create what synda does in R cause a do X in CLI then Y in R isn't ideal - i'll play around - will probably need to put this review on hold for a while 😸
