### 2017-06-19 #99 @sckott
now that `charlatan` is through review - i'll return to this one - it still may be a few weeks or so - reviewer pointed out that we should really be leveraging much more of the data available, so working on that soon - not a clear easy path with the data available though :(
