### 2017-07-25 #99 @maelle
@sckott any progress?

In any case you can add the peer-review badge to `cmipr` README via
