### 2017-07-25 #99 @sckott
@maelle will do the badge

blarg, no sorry, no progress. will try to get to very soon.
