### 2017-10-02 #99 @sckott
sorry @maelle about the delay!  I have on to do list to look into the new data source options, and have not done so yet. will try to get it done asap
