### 2018-07-13 #99 @sckott
@bpbond Finally trying to get back to this. Looking at [api][] and the python client http://esgf-pyclient.readthedocs.io/en/latest/ and  trying the same in R (been using base url https://esgf-index1.ceda.ac.uk so far), but it seems like I can only use the search API freely, but many data files I've tried so far are behind login walls. Do you know what it takes to get access to ESGF data?

Is it a valuable contribution to provide an R interface to the [search API][api] at least?

[api]: https://github.com/ESGF/esgf.github.io/wiki/ESGF_Search_REST_API
