### 2018-07-13 #99 @bpbond
Hey @sckott . That's true as far as I know--downloading data requires an ESGF login account; note that ESGF is a source of huge [frustration](https://twitter.com/DustyBowl/status/1016658859279245312) to many researchers.

I think even exposing the search API through R would be valuable, yes! Anything to make the system easier to use...
