### 2018-07-13 #99 @sckott
Thanks @bpbond - Okay, sounds like best approach moving forward is to just give access to any freely available cmip data I can, and have a search API interface. Then if folks with access to the actual data files want to contribute later on that's great
