### 2018-11-09 #99 @sckott
https://github.com/ropenscilabs/cmipr#cmipr
